+++ bindings/nostr-sdk-ffi/src/protocol/nips/nip46.rs
@@ -2,6 +2,7 @@
 // Copyright (c) 2023-2024 Rust Nostr Developers
 // Distributed under the MIT software license
 
+use std::fmt;
 use std::ops::Deref;
 use std::str::FromStr;
 use std::sync::Arc;
@@ -158,9 +159,15 @@ impl NostrConnectMetadata {
 }
 
 #[derive(Debug, PartialEq, Eq, Hash, Object)]
-#[uniffi::export(Debug, Eq, Hash)]
+#[uniffi::export(Debug, Display, Eq, Hash)]
 pub struct NostrConnectURI {
     inner: nip46::NostrConnectURI,
+}
+
+impl fmt::Display for NostrConnectURI {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{}", self.inner)
+    }
 }
 
 impl From<nip46::NostrConnectURI> for NostrConnectURI {
@@ -184,10 +191,6 @@ impl NostrConnectURI {
         Ok(Self {
             inner: nip46::NostrConnectURI::parse(uri)?,
         })
-    }
-
-    pub fn as_string(&self) -> String {
-        self.inner.to_string()
     }
 }